+++ TATPhysics/TATResources/TATObjLoader.h
@@ -54,7 +54,7 @@ public:
 			{
 				std::vector<int> indexs;
 				int size = 0;
-				for (int i = 1; i < splitResult.size(); i++)
+				for (int i = 1; i < (int)splitResult.size(); i++)
 				{
 					std::vector<TString> datas = splitResult[i].Split("/");
 					size = datas.size();
@@ -75,7 +75,7 @@ public:
 					}
 
 				}
-				if (indexs.size() > 3 * size)
+				if ((int)indexs.size() > 3 * size)
 				{
 					for (int i = 0; i < size; i++)
 					{
@@ -86,7 +86,7 @@ public:
 						indexs.push_back(indexs[2 * size + i]);
 					}
 				}
-				for (int i = 0; i < indexs.size() / (3*size); i++)
+				for (int i = 0; i < (int)indexs.size() / (3*size); i++)
 				{
 					if (size == 1)
 						buffer.faceBuffer.push_back(TATFaceBuffer(indexs[i * 3 * size + 0 * size], indexs[i * 3 * size + 1 * size], indexs[i * 3 * size + 2 * size]));
@@ -106,13 +106,13 @@ public:
 			else if (splitResult[0] == "t")
 			{
 				std::vector<int> indexs;
-				for (int i = 1; i < splitResult.size(); i++)
+				for (int i = 1; i < (int)splitResult.size(); i++)
 				{
 					std::vector<TString> datas = splitResult[i].Split("/");
 					int index = datas[0].ToInt();
 					indexs.push_back(index - 1);
 				}
-				for (int i = 0; i < indexs.size() / 4; i++)
+				for (int i = 0; i < (int)indexs.size() / 4; i++)
 				{
 					buffer.tetraBuffer.push_back(TATetraBuffer(indexs[i + 0], indexs[i + 1], indexs[i + 2], indexs[i + 3]));
 				}
@@ -121,7 +121,7 @@ public:
 
 		if (buffer.normalBuffer.size()>0)
 		{
-			for (int i = 0; i < buffer.faceBuffer.size(); i++)
+			for (int i = 0; i < (int)buffer.faceBuffer.size(); i++)
 			{
 				for (int c = 0; c < 3; c++)
 				{
